VictimProjects/WeatherStation/web_dashboard/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
paho-mqtt==2.1.0
gunicorn==21.2.0
